⚡ :: (#32) 흩어져 있는 타입들을 types.ts에 모아놈
