+++ volvo/src/constance/SelectType/index.ts
@@ -1,5 +1,3 @@
-export const SCHOOL = '대덕소프트웨어마이스터고등학교';
-export const SELECTTYPE = '2022년 지원자 전형 구분';
 export type typeInfoIdType = 'regular' | 'meister';
 export type typeInfoType = {
   id: typeInfoIdType;
@@ -56,31 +54,3 @@ export const HEADCOUNT: Array<headCountType> = [
 ];
 export const GRADUATION_DATE_EXPLAIN = '졸업 예정자의 경우 졸업 예정월만 선택해주세요';
 export const REMARK_EXPLAIN = '해당하는 특기사항에 체크해주세요';
-
-export type selectTypeType = {
-  educational_status: string | null;
-  application_type: string | null;
-  is_daejeon: boolean | null;
-  application_remark: string | null;
-  graduated_at: string | null;
-  headcount: string | null;
-};
-
-export type selectTypeSaveType = {
-  educationalStatus: string | null;
-  applicationType: string | null;
-  isDaejeon: boolean | null;
-  applicationRemark: string | null;
-  graduationMonth: number | null;
-  graduationYear: number | null;
-  headcount: string | null;
-};
-
-export interface selectTypeInterface {
-  educational_status: string | null;
-  application_type: string | null;
-  is_daejeon: boolean | null;
-  application_remark: string | null;
-  graduated_at: string | null;
-  headcount: string | null;
-}
