+++ volvo/src/containers/grade/index.tsx
@@ -1,13 +1,11 @@
-import Grade from "../../components/grade";
-import { useGrade } from "../../util/hooks/grade";
+import React from 'react';
+import Grade from '../../components/grade';
+import { useGrade } from '../../util/hooks/grade';
 
+const GradeContainers = () => {
+  const { state, setState } = useGrade();
 
-const GradeContainers =()=>{
-    const {state,setState}=useGrade();
-    
-    return(
-        <Grade {...state} {...setState}/>
-    )
-}
+  return <Grade {...state} {...setState} />;
+};
 
 export default GradeContainers;
