--- volvo/src/models/dto/request/signinRequest.ts
@@ -1,4 +1,0 @@
-export interface signinRequest {
-  email: string;
-  password: string;
-}
