--- volvo/src/models/dto/request/statusRequest.ts
@@ -1,1 +1,0 @@
-export interface statusRequest {}
