--- volvo/src/models/dto/request/userRequest.ts
@@ -1,1 +1,0 @@
-export interface userRequest {}
