--- volvo/src/models/dto/response/responseGenerator.ts
@@ -1,8 +1,0 @@
-export interface responseGenerator {
-  config?: any;
-  data?: any;
-  headers?: any;
-  request?: any;
-  status?: number;
-  statusText?: string;
-}
