--- volvo/src/models/dto/response/signinResponse.ts
@@ -1,8 +1,0 @@
-export interface signinResponse {
-  access_token: string;
-  refresh_token: string;
-}
-
-export interface refreshResponse {
-  access_token: string;
-}
