--- volvo/src/models/dto/response/signupResponse.ts
@@ -1,16 +1,0 @@
-export interface signupResponse {
-  access_token: string;
-  refresh_token: string;
-}
-
-export interface checkVertifyResponse {}
-
-export interface signupVertifyCodeResponse {}
-
-export interface resetPasswordVertifyCodeResponse {}
-
-export interface resetPasswordResponse {}
-
-export interface signUpResponse {
-  access_token: string;
-}
