--- volvo/src/models/dto/response/statusResponse.ts
@@ -1,4 +1,0 @@
-export interface statusResponse {
-  schedules: { type: string; date: string; process: string }[];
-  current_status: string;
-}
