--- volvo/src/models/error.ts
@@ -1,5 +1,0 @@
-export type error = {
-  status: number;
-  message: string;
-  type: string;
-};
