+++ volvo/src/modules/redux/reducer/preview/interface/index.ts
@@ -1,4 +1,4 @@
-import { error } from '../../../../../models/error';
+import { error } from '../../../../../types';
 
 interface PreviewState {
   preview: string;
