+++ volvo/src/modules/redux/reducer/signin/interface/index.ts
@@ -1,4 +1,4 @@
-import { error } from '../../../../../models/error';
+import { error } from '../../../../../types';
 
 interface ISignInState {
   id: string;
