+++ volvo/src/util/api/status/index.ts
@@ -1,9 +1,8 @@
 import uri from '../../../constance/uri';
-import { statusRequest } from '../../../models/dto/request/statusRequest';
-import { statusResponse } from '../../../models/dto/response/statusResponse';
+import { statusResponse } from '../../../types';
 import { getRequest } from '../default';
 
-export const getStatus = async (statusRequest: statusRequest) => {
+export const getStatus = async () => {
   const request = getRequest();
   const response = await request.get<statusResponse>(uri.schedule);
   return response;
